refactor: make parent citation service
