+++ app/components/citation_component.html.erb
@@ -1,12 +1,26 @@
 <div class="citation-details">
   <h2 class="h4"><%= t("citation.persistent_id") %></h2>
   <div data-controller="clipboard" data-clipboard-success-content-value="<%= t("citation.copied") %>">
-    <p><%= link_to "https://nla.gov.au/nla.cat-vn#{@document.id}", "https://nla.gov.au/nla.cat-vn#{@document.id}", { data: { "clipboard-target": "source" } }
-    %> <button type="button" data-action="clipboard#copy" data-clipboard-target="button" class="btn btn-icon" aria-label="Copy to clipboard"><i class="fa-regular fa-copy" aria-hidden="true"></i></button></p>
+    <p><%= link_to "https://nla.gov.au/nla.cat-vn#{@document.id}", "https://nla.gov.au/nla.cat-vn#{@document.id}", { data: { "clipboard-target": "source" } } %>
+      <a href="#" data-action="clipboard#copy" data-clipboard-target="button" title="<%= t("citation.copy_to_clipboard") %>" class="text-decoration-none ml-1"><i class="fa-regular fa-copy" aria-hidden="true"></i></a></p>
   </div>
 
   <% @formats.each do |i18n_key, citation_method| %>
-    <h2 class="h4"><%= t(i18n_key) %></h2>
-    <p><%= @document.send(citation_method).html_safe %></p>
+    <% if citation_method != :export_as_nla_wikipedia_citation_txt %>
+      <h2 class="h4"><%= t(i18n_key) %></h2>
+      <div data-controller="clipboard" data-clipboard-success-content-value="<%= t("citation.copied") %>">
+        <p><%= @document.send(citation_method).html_safe %>
+          <a href="#" data-action="clipboard#copy" data-clipboard-target="button" title="<%= t("citation.copy_to_clipboard") %>" class="text-decoration-none ml-1"><i class="fa-regular fa-copy" aria-hidden="true"></i></a></p>
+      </div>
+    <% end %>
   <% end %>
+
+  <h2 class="h4"><%= t("citation.wikipedia.title") %></h2>
+  <%= t("citation.wikipedia.help").html_safe %>
+  <div class="card">
+    <div class="card-body" data-controller="clipboard overflow-auto" data-clipboard-success-content-value="<%= t("citation.copied") %>">
+      <pre data-clipboard-target="source" class="p-1 m-0"><%= @document.send(:export_as_nla_wikipedia_citation_txt).html_safe %></pre>
+      <a href="#" data-action="clipboard#copy" data-clipboard-target="button" title="<%= t("citation.copy_to_clipboard") %>" class="text-decoration-none position-absolute" style="top: 1rem; right: 1rem;"><i class="fa-regular fa-copy" aria-hidden="true"></i></a>
+    </div>
+  </div>
 </div>
