+++ spec/files/marc/3601830.marcxml
@@ -1,0 +1,1 @@
+<record xmlns='http://www.loc.gov/MARC21/slim'><leader>00607pam a2200229 a 4500</leader><controlfield tag='001'>3601830</controlfield><controlfield tag='005'>20171211152048.0</controlfield><controlfield tag='008'>051213s2006    xnaa          001 0 eng  </controlfield><datafield ind1=' ' ind2=' ' tag='020'><subfield code='a'>9781740457590 :</subfield><subfield code='c'>$29.95</subfield></datafield><datafield ind1=' ' ind2=' ' tag='020'><subfield code='a'>1740457595</subfield></datafield><datafield ind1=' ' ind2=' ' tag='035'><subfield code='a'>3601830</subfield></datafield><datafield ind1=' ' ind2=' ' tag='040'><subfield code='a'>ANL</subfield><subfield code='b'>eng</subfield><subfield code='d'>ANL</subfield></datafield><datafield ind1=' ' ind2=' ' tag='042'><subfield code='a'>anuc</subfield></datafield><datafield ind1='0' ind2='4' tag='082'><subfield code='a'>746.432</subfield><subfield code='2'>22</subfield></datafield><datafield ind1='0' ind2='0' tag='245'><subfield code='a'>Knit /</subfield><subfield code='c'>designs by Diana Crossing ... [et al.]</subfield></datafield><datafield ind1=' ' ind2=' ' tag='260'><subfield code='a'>Sydney :</subfield><subfield code='b'>Murdoch,</subfield><subfield code='c'>2006.</subfield></datafield><datafield ind1=' ' ind2=' ' tag='300'><subfield code='a'>111 p. :</subfield><subfield code='b'>col. ill. ;</subfield><subfield code='c'>25 x 23 cm.</subfield></datafield><datafield ind1='1' ind2=' ' tag='490'><subfield code='a'>Handmade style</subfield></datafield><datafield ind1=' ' ind2=' ' tag='500'><subfield code='a'>Includes index.</subfield></datafield><datafield ind1=' ' ind2='0' tag='650'><subfield code='a'>Knitting.</subfield></datafield><datafield ind1='1' ind2=' ' tag='700'><subfield code='a'>Crossing Diana.</subfield></datafield><datafield ind1=' ' ind2='0' tag='830'><subfield code='a'>Handmade style.</subfield></datafield><datafield ind1='f' ind2='f' tag='999'><subfield code='i'>8f08941e-3689-5f1a-b969-0f0541f0e0dd</subfield><subfield code='s'>435c1bf3-675a-59d6-b172-749dd5487211</subfield></datafield><datafield ind1='f' ind2='f' tag='952'><subfield code='a'>National Library of Australia</subfield><subfield code='b'>Parkes</subfield><subfield code='c'>Parkes</subfield><subfield code='d'>Main Reading Room (AUS [Lakeside])</subfield><subfield code='e'>746.432 K71</subfield><subfield code='h'>Dewey Decimal classification</subfield><subfield code='i'>Monograph</subfield><subfield code='m'>a5b51891-b583-5113-8a32-a57ea7670fee</subfield><subfield code='z'>Holdings note: N pbk</subfield></datafield></record>
