+++ spec/files/marc/8425632.marcxml
@@ -1,0 +1,1 @@
+<record xmlns='http://www.loc.gov/MARC21/slim'><leader>01751cas a2200349 a 4500</leader><controlfield tag='001'>8425632</controlfield><controlfield tag='005'>20200722144547.0</controlfield><controlfield tag='006'>m x p|o     0   a0</controlfield><controlfield tag='007'>cr mn </controlfield><controlfield tag='008'>130528d20033003cc  x p|o     0   a0chi d</controlfield><datafield ind1=' ' ind2=' ' tag='035'><subfield code='a'>(CNKI)CAJ201213156</subfield></datafield><datafield ind1=' ' ind2=' ' tag='040'><subfield code='a'>CNKI</subfield><subfield code='b'>eng</subfield></datafield><datafield ind1=' ' ind2=' ' tag='130'><subfield code='6'>880-01</subfield><subfield code='a'>Zhui xun de xuan lv (Online)</subfield></datafield><datafield ind1='1' ind2='0' tag='245'><subfield code='6'>880-02</subfield><subfield code='a'>Zhui xun de xuan lv</subfield><subfield code='h'>[electronic resource].</subfield></datafield><datafield ind1=' ' ind2='1' tag='260'><subfield code='6'>880-03</subfield><subfield code='a'>Nanjing :</subfield><subfield code='b'>Jiangsu Sheng Tao Xingzhi yan jiu hui,</subfield><subfield code='c'>2003-3003.</subfield></datafield><datafield ind1=' ' ind2=' ' tag='310'><subfield code='a'>No determinable frequency</subfield><subfield code='b'>2003-2003</subfield></datafield><datafield ind1='1' ind2=' ' tag='362'><subfield code='a'>Began in 2003;ceased in 2003.</subfield></datafield><datafield ind1=' ' ind2=' ' tag='500'><subfield code='a'>Title from cover.</subfield></datafield><datafield ind1=' ' ind2=' ' tag='533'><subfield code='a'>Electronic reproduction.</subfield><subfield code='b'>Beijing :</subfield><subfield code='c'>China academic journals electronic publishing house.</subfield></datafield><datafield ind1=' ' ind2=' ' tag='538'><subfield code='a'>System requirements: Windows 98 above; IE browser is recommended; CAJ viewer is required to view full text or Acrobat Reader with Adobe simplified.</subfield></datafield><datafield ind1=' ' ind2='0' tag='650'><subfield code='a'>Secondary Education</subfield><subfield code='v'>Periodicals.</subfield></datafield><datafield ind1='2' ind2=' ' tag='710'><subfield code='6'>880-04</subfield><subfield code='a'>Jiangsu Sheng Tao Xingzhi yan jiu hui.</subfield></datafield><datafield ind1='0' ind2=' ' tag='740'><subfield code='a'>China Academic Journals - Education &amp; Social Sciences (Series H) - Chinese</subfield></datafield><datafield ind1='1' ind2=' ' tag='773'><subfield code='t'>China academic journals</subfield></datafield><datafield ind1='0' ind2='0' tag='780'><subfield code='6'>880-05</subfield><subfield code='t'>Xin shi dai de jiao bu sheng</subfield></datafield><datafield ind1='8' ind2=' ' tag='852'><subfield code='b'>ELECASI </subfield><subfield code='h'>Electronic resource </subfield><subfield code='x'>Available for document delivery - restrictions apply </subfield><subfield code='z'>Available to National Library Registered Users.</subfield></datafield><datafield ind1='4' ind2='0' tag='856'><subfield code='3'>China academic journals</subfield><subfield code='u'>https://oversea.cnki.net/knavi/JournalDetail?pcode=CJFD&amp;pykm=SDDB</subfield><subfield code='z'>Connect to periodicals (National Library of Australia only)</subfield></datafield><datafield ind1='1' ind2=' ' tag='880'><subfield code='6'>130-01</subfield><subfield code='a'>追寻的旋律(Online)</subfield></datafield><datafield ind1='1' ind2='0' tag='880'><subfield code='6'>245-02</subfield><subfield code='a'>追寻的旋律</subfield><subfield code='h'>[electronic resource].</subfield></datafield><datafield ind1=' ' ind2='1' tag='880'><subfield code='6'>260-03</subfield><subfield code='a'>南京 :</subfield><subfield code='b'>江苏省陶行知研究会,</subfield><subfield code='c'>2003-3003.</subfield></datafield><datafield ind1='2' ind2=' ' tag='880'><subfield code='6'>710-04</subfield><subfield code='a'>江苏省陶行知研究会.</subfield></datafield><datafield ind1='0' ind2='0' tag='880'><subfield code='6'>780-05</subfield><subfield code='t'>新时代的脚步声</subfield></datafield><datafield ind1='f' ind2='f' tag='999'><subfield code='i'>1a665106-83d3-5f7a-a46e-6c679bae6e71</subfield><subfield code='s'>d023e48c-c5cb-5e97-b7e2-733bf725bd46</subfield></datafield><datafield ind1='f' ind2='f' tag='952'><subfield code='a'>National Library of Australia</subfield><subfield code='b'>Parkes</subfield><subfield code='c'>Parkes</subfield><subfield code='d'>Use &apos;Online Access&apos; link in record (ELECASI)</subfield><subfield code='e'>Electronic resource </subfield><subfield code='h'>Other scheme</subfield><subfield code='i'>Serial</subfield><subfield code='m'>233e7bbf-8433-55b0-bbff-5e50d7316774</subfield></datafield></record>
